+++ ChangeLog.md
@@ -15,8 +15,13 @@
 
 - Add new plugin `MultiClient` and new flag `--multiclientscript` to handle package structure of multi client  #1328
 
+**Bug Fixes**
+
+- Fix regression in default namespace for SDKs generated without `--namespace` flag  #1354
+
 **Other Changes**
 
+- Generated code no longer supports Python 3.6  #1353
 - Order json input and response template entries by whether they are required or not #1335
 - Reduce extreme amount of `black` logs when running in `--debug` mode to just log errors